requirements.txt
Django==1.10.4
django-crispy-forms==1.6.1
django-registration-redux==1.4
